requirements.txt
Telethon==1.40.0
python-dotenv==1.1.1
pytest==8.4.1
pytest-asyncio==1.1.0
twilio==9.7.0
requests==2.31.0
aiohttp==3.9.5
Pillow==10.3.0
